[Bronze V] Title: 이상한 기호, Time: 212 ms, Memory: 17696 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -5,10 +5,11 @@ public class Main {
 	public static void main(String[] args) {
 		Scanner sc= new Scanner(System.in);
 		
-		int a=sc.nextInt();
-		int b=sc.nextInt();
+		long a=sc.nextLong();
+		long b=sc.nextLong();
 		
-		int ab=((a+b)*(a-b));
+		long ab=((a+b)*(a-b));
 		System.out.println(ab);
+		
 	}
 }
